refactor and fix signal reconstruction
